+++ capi/example/soem.c
@@ -4,7 +4,7 @@
  * Created Date: 16/05/2022
  * Author: Shun Suzuki
  * -----
- * Last Modified: 01/11/2022
+ * Last Modified: 15/11/2022
  * Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
  * -----
  * Copyright (c) 2022 Shun Suzuki. All rights reserved.
@@ -71,7 +71,7 @@ int main() {
     return -1;
   }
 
-  AUTDSetCheckTrials(cnt, 50);
+  AUTDSetAckCheckTimeout(cnt, 20 * 1000 * 1000);
 
   return run(cnt);
 }